Add translation library #31
This commit adds transloco as a translation library for Angular. Some
text has been replaced with translation keys to show the concept of how
the library can be used.
This commit also adds transloco-keys-manager to extract and check the
use of translation keys. The check command also has been added to the
pre-commit hook to help finding missing or extraneous keys.

The constant defines what languages are supported by our app. This
constant is used to create a type that describes valid language
identifiers.

The type definition of the transloco configuration requires a mutable
type to be provided for availableLangs property. As we use 'as const' in
our constant definition, TypeScript rejects our array as a readonly type
cannot be assigned to a mutable type. To solve this we create a copy of
our language definitions.

Issue #31