Add translation library
This commit adds transloco as a translation library for Angular. Some
text has been replaced with translation keys to show the concept of how
the library can be used.
This commit also adds transloco-keys-manager to extract and check the
use of translation keys. The check command also has been added to the
pre-commit hook to help finding missing keys or extraneous keys.

Part of issue #5