Add github workflow to deploy to github pages
This commit adds github workflows to test and deploy the application to
github pages. There are two workflows, one for (manual) deployment and
one for PRs. The PR workflow also triggers a deployment, if the PR is
merged into the main branch.

Some fixes also had to be applied to make the workflow work. For
example, the linter configuration is an ES-module and tried to import a
folder which does not work for ES-modules but only works for CommonJS
modules.

Issue #4