Add github workflow to deploy to github pages #4
This commit adds github workflows to test and deploy the application to
github pages.

The idea is to deploy the main branch automatically when a PR is merged
or a commit is pushed to main. This gives an up-tp date preview of the
current state of the application. This can be used to show the current
state to stakeholders.

Deployment should also be possible manually for feature branches. This
is why the deployment code is in its own workflow to allow manual
trigger.

Then there is the workflow for code-quality that is used for PRs. It
performs different checks that should pass. The different checks are put
into individual jobs on purpose. This causes github to execute the jobs
in parallel and allows you to see the results of all jobs directly in
the PR. This means for example if the formatting does not pass you can
still also see the results of the tests. If at some point there are too
many checks some can be consolidated.

Some fixes also had to be applied to make the workflow work. For
example, the linter configuration is an ES-module and tried to import a
folder which does not work for ES-modules but only works for CommonJS
modules.

Issue #4