Add language types and constants
The constant defines what languages are supported by our app. This
constant is used to create a type that describes valid language
identifiers.

The type definition of the transloco configuration reuqires a mutable
type to be provided for availableLangs property. As we use 'as const' in
our constant definition TypeSciript rejects our array. To solve this we
create a copy of our language definitions.

Issue #4