Fix fetch of translation files not using base-href #31
The JSON files for the translation service are fetched using a
HTTPClient. Angular cannot detect that this HTTPClient is doing a
request to public assets and thus cannot prepend the base-href
automatically. Angular does provide a function in the Location service
to generate a URL that uses the base-href prefix which is used in this
commit.