Fix auto-deploy workflow executed twice when a PR is merged #4
The auto-deploy workflow is intended to automatically deploy the main
branch to github-pages whenever a PR is merged or something is pushed to
the main branch. However, merging a PR apparently also pushes to the
main branch which causes the workflow to be triggered twice, once for
closing the PR and once for pushing to the main branch. This means,
however, that the trigger on closed PRs is redundant and can be removed.