Add service to query parameters from STAC API #47
The Stac-api service uses a generated APICLient to access the STAC-API
service. The APIClient is generated using the OpenAPI definition of the
STAC API.

For each collection within the STAC-API there are three metadata files.
One for parameters, one for stations and one that maps stations to
parameters. Each meta file will be managed by its own service who is
also responsible to transform the CSV data into an internal
representation. This commit contains the service for parameters.

For a id of a parameter the last two digits are excluded. This is done
because those two digit represent the type and interval of the
measurement. So there are multiple parameters with the same prefix in
the shortname-key but with different postfix. For our application this
distinguishing is not needed, which is why we exclude it so the
parameters are treated as the same parameter rather than multiple.