Put loaded parameters into ngrx store #47
Parameters are loaded asynchronously for multiple collections and once
loaded put into the store for the rest of the application to access it.
The design is kept simple for now, so that a single error for one
collection will be interpreted as a error for the whole process. This
can be handled more granular later.

In order to facilitate debugging the store the package
@ngrx/store-devtools has been added. This allows the browser-extension
"Redux Dev Tools" to see events on the store and show how actions affect
the store.