Sort typescript imports #32
Sorting imports makes the code more consistent and it is easier to
distinguish between library imports and local imports. This is
especially useful if there are a lot of imports in a single file. The
plugin is configured to distinguish between type and value imports and
it will merge multiple imports form the same source.

To do this, this commit adds the prettier plugin package
"@ianvs/prettier-plugin-sort-imports" and configures it accordingly.

Prettier is used rather than eslint to sort imports the order of imports
is considered to have no influence on semantics of the code. Imports
should not have side-effects. The plugin would consider side-effects
imports and make sure not to change the semantics. However, in Angular
these are not common anyway.