Add prettier-ignore statement to group imports in eslint config #32
The imports are grouped in a way such that the comment can explain why
these imports should be revised in the future. So the import plugin
should not move the import statements to the other group. This can be
done by adding a prettier-ignore comment above the import group which
prevents the plugin from sorting imports away.