Add station service #56
This service loads all available stations for the given collections and
returns the merged values. It is written analogously to the parameter
service. The effects for the station store fetch the stations from the
service and put it in the store. The service transforms the CSV entries
to the internal representation.

When loading a collection both the parameters and the stations have to
be loaded all the time. This is now done by using a collection action
that triggers both effects for station and parameters.