Refactor station and parameter service #56 [PR]
The CSV interface definitions are moved to their own files to make the
service cleaner.

Some methods of the services are made private. This requires a refactor
of the tests too, as the type system complains when spying on private
methods. Thus the StacApiService is mocked and retunes CsvStation data.
As this implies that every test now tests the transform method an
example is put on top of the file. Each test now uses this and modifies
only the required properties, such that changes to the transformation
does not break other tests.