Fix error handling and added it for stations #32
The error handling as we had it until now was broken. The original error
(the cause) was missing from all our errors. The unit tests that should
have detected this had also some issue with deep equality.
I fixed both in this commit and also added error handling for stations.