Add measurement data type #55
The parameters and stations should be selected based on a selected
measurement data type. Each measurement data type defines a set of
collections that are loaded with their own stations and parameters. The
form behaves the same for each measurement type, except that the data of
each measurement type is not mixed with the others.

Changing the measurement data type selection loads the corresponding
collections, if they are not loaded yet.