Refactor effects registration #32
Effects need to be registered in ngrx for them to take effect. We did
this by defining an array where all effects are put in and this array is
given to ngrx. ngrx expects the contents of the array to be objects
containing effect functions not the functions themselves, which is why
we had objects wrapping the different effects in the array.

However, we tend to forget to register the effects. This can lead to
unintended effects, like an error handler not triggering. When using
Effects classes this problem is solved by providing the class to ngrx
which contains all effects. However, we are using functional effects so
we have to do that a bit differently.

To make effects automatically registered, we import all exported members
of the effects file using a * import and alias it to a meaningful name.
This allows us to have an object containing all effects, as the imported
module acts like an object in JavaScript terms. This way we can just add
the aliased imports to the array and don't need to worry about
forgetting an import.

However, this means that effects files cannot export anything else but
effects. But they should not need to do that anyway. In the worst case
we have to either filter the members of the import for effects or
specify the imports manually and alias them.