Add search input for stations #58
Station names can be searched using Material's autocomplete component
There are two pipes necessary: One listening to the form changing and
pushing that to the store and one listening to the store doing the same
the other way around. However, there is no endless loop happening due to
the fact that the form only emits **changing** values.