Add rule to require comment for eslint-disable instructions #32
Eslint rules are set up to ensure a consistent code quality. Disabling
rules manually can cause a degradation in code quality. However, there
are some rare situations, where eslint rules need to be disabled. To
ensure proper documentation a eslint plugin is added that enforces
comments to describe the reason why the rule has been disabled. The
plugin is chosen over a convention, because if a convention is not
enforced by tooling, it is likely to be forgotten in the future.