Add collections to data structures #66
To be able to select the correct feature data for the selection the user
made, we need to know which collection the requested data is in. To
track this, the collection is added to stations and station-parameter
mappings.

The collection is added to the station as we need to get the feature for
a specific station. We track all unique collections the station is in, if
there are multiple.

In the end we need to narrow down the selection to a single collection,
so we can guide the user to the searched for data. We intent to filter
those collections using the selected parameter group, if selected, or
else force the user to decide.

To filter down the stations collections by the selected parameter group
we need to know which parameters are in which collection. However, we
don't actually need this information for the parameter itself, but
always in combination with the station, which is why we track that
information on the parameter-station mapping data structure.

As the user only selects a parameter group and not a specific parameter,
we also track the collections in the station-parameter-group-mapping.

This commit also improves the test specifications for the station-,
parameter- and station-parameter-mapping-service. They have been
expanded to include the properties about collections and how only unique
collections should be tracked.

The test for testing if the transform works correctly has been removed,
for the station and parameter-service, as it is an internal
implementation detail on how it works. Its behavior can be observed by
using the public `load<whatever>` method. This means that tests are less
likely to break if that method is changed.

Further, the description property of the parameter type is removed, as
it is not used anywhere and we don't intent to use it as of now.