Add collection selection step #66
This step is used to narrow down the collections the station is in to a
single collection. This allows us to make sure the user only sees
available selection options in the later steps, as we know which feature
to query from the stac-API. The intention is to skip this step, if
already only a single collection can be considered due to station
selection or parameter group selection. This is not implemented in this
commit yet.