Enable in-/export of app state from/to URL #53
`route-store` is an additional package for NGRX to grant easy access to
the Angular Router state. It incorporates reading it as well as access
to common actions like navigation start/end.

The application can now be initialized using the current URL fragments.
If the fragments don't contain the necessary information the default
values will be taken instead. This whole process is controlled in the
app state. There is also a new property `isInitialized` indicates
whether this initialization is already done (preventing endless loops of
trying to apply the newest URL fragments to the app state).
Each state that is interested in these inital values can install an
effect listening to the `appActions.initializeApp` action and handle
accordingly - this is usually done by calling the next action to set the
desired value.

`url-parameter.service` was created to encapsulate the communication
between our state and the URL parameters. All currently used parameters
are stored within the new `app-url-parameter` model, so that other parts
of the application can easily pick the necessary information.