Fix station add/filter flow #34
The order in which the stations were added to the map and filtered was
not defined previously. This could result in an error when a filter was
applied to the map where no layers are available.
This was solved using a new state property `areLayersInitialized` which
marks the map as fully initialized.