Show station type as name for collection selection #66
The collection identifier is not very insightful for users. Further,
STAC does not provide a translated name for collections. However, the
station csv contains the station_type, which is the same for all
stations within a collection, so we can use this as a representation of
a collection name.

To make it easier to track collection id and stationType for each
station, the station type is adapted to keep the collection id and the
station type and we no longer merge them when reading them from the CSV.
This requires some adaption in the map, so we only show the most
important station on the map.