Display known parameter groups for available station type #66
To make an informed decision which station type to select, the parameter
groups contained in the station type should be displayed. For this the
forms selector need to be adapted to group stations with their
corresponding parameter groups.

If only a single station type is available, either because the station
only has one, or the selected parameter group already filters it out, we
can auto-select the corresponding selection. This is done by using an
effect that listens on the selectStationId action.