Add service to fetch assets form STAC #41
This service retrieves all assets for a given station and collection
from the stac API. As only the asset filenames and urls are needed we
only use these values. Then we extract some information from the
filenames for which the format is predefined.

The errors in this service are inherited from `SilentError`
Failing to load an asset should not crash the application in contrast to
our previous errors while loading the stations, parameters or their
mapping which is essential to the application.