Combine URL param set into one effect/method #53
There are three different (but interwoven) changes:
* url-parameter.service is heavily simplified by only accepting the new
   app-url-parameter and setting them
* app-url-parameter contains now all form state properties
* collection is now part of the URL and the initialization

The url-parameter.effects are now responsible to listen for the right
actions and set the URL parameter accordingly. This makes it much easier
to add new parameters as there is only one place to modify.

Also there was a minor issue in station-selection.component where the
station ID was dispatched too many times even if nothing else changed.

The app URL state should be complete now with these changes. The only
thing missing is the initialization of the data interval and time range
which is done in their own PR.