Add interval selection #60
The interval selection allows a user to select which interval the
requested data should have. The interval selection always shows all
possible intervals but only allows selecting those that are available
given the station assets.

The labels for the possible interval are mapped in an object in the
component. This allows transloco key manager to find them using the
marker function even though the html is generated based on all
existing intervals.