Add assets to state #41
Assets are now loaded automatically when a collection is set because at
that point we have all necessary information (station ID + collection).
The effect for loading the assets is slightly different than our usual
approach: it loads new assets independent from the loading state. It
does not matter if any other assets are already loaded.

Also any change to station ID or collection will either result in a
new request or the state will be reset as it heavily depends on the
currently selected station and collection.

Co-authored-by: Sebastian Wendland <5629793+wendlans@users.noreply.github.com>