Add time range selection #61
This selection allows a user to select which time range the requested
data should have. Compared to interval selection time-range is a bit
special. Due to limits in excel line counts, the historical data is
can be split into junks. This is not always the case though. If,
however, such assets exist all historical assets without it are ignored.

Due to date ranges now being a thing, the form state is extended with
to store the selected daterange. This is done to make sure the asset
selection can be distinct.