Put collection assets in store #32
Similar to stations a collection can contain assets. In fact for our
application it is crucial to retrieve them as they contain the links to
the meta CSV files that is needed to display all necessary values. As
there are multiple parts of the application that depend on this
information we put the collection assets into the store so they can be
used again.