Add MeteoSchweiz parameter group icons #48
IconService is used to register all svg items from a config to Material.
There are two main reasons to use this service: it simplifies the
handling of local URLs and secondly it enforces us to list all icons in
config files using defined IDs (independent from the file name). So if
anything changes we only have to change this config.