Add pipe to test if value is included in a container #61 [PR]
To test whether we should allow selection of certain items we look if
the value is contained in the list of available values. To prevent
having to call .includes in the template a pipe is added. This pipe
allows to search for the values in any container that implements
`includes`, like string or arrays, or implements `has`, like Map and
Set. Additionally null and undefined are allowed to allow this pipe to
be used directly after an async pipe which might return null. The
searched value is typed to be assignable to the container it is looked
for to help with type safety.