Replace parameter group selection using autocomplete #48
Parameter groups can now be search/selected in the same way as stations.
The only difference is that the parameter list stays unfiltered if an
item is selected.
Parameter group also got a new inherited interface
`LocalizedParameterGroup` which contains the translated name of the
group as it easier and more efficient to translate that in the selector.