Fix app initialization #53
There is a distinct init-action for each initial parameter to make sure
that there is no loop back to setting the URL while the app is not yet
finished initializing. These actions don't change the state apart from
their value.

Also the default value from `data-selection-form.component` was removed.
This is set within the app initialization using the URL fragment or
default values if the fragment is empty/invalid.