Initialize data interval and time range from URL #53
With this it's possible to complete restore the form state using the
initial URL parameters.
The tricky part in this commit was to get the timing right:
First the collections must be loaded then the station+collection
verified, the corresponding assets loaded and so that finally the data
interval and (historical) time range can be verified using the assets.