Use BEM and general design clean up #48
This change introduces BEM to structure the SCSS files. Firstly, it
makes it easier to identify the corresponding class usage in the
templates (e.g. using the CTRL-click) as it is unique afterward. And
secondly, it makes refactorings (especially renamings) a lot less
dangerous as the IDE sometimes does rename a lot more than necessary.

The whole parameter group and station selection part is now a lot closer
to the desired design. The debug options to switch the language and
throw an error is removed.