Style measurement-data-type-selection #86
Since the measurement-data-type selection is now bigger than just a
button-toggle-group it has been moved into its own component to keep the
data-selection-form component clean. In that new component the info box
for the measurement-data-type is added.

This change causes an issue in that the form is no longer reset, if the
measurement-data-type is changed, as the stepper can no longer be
accessed from the toggle-button group. However, the implementation of
resetting the stepper was hacky anyway.

In the ticket #87 the logic of the stepper should be revised and this
change can include reintroducing this feature again as well. Using
<Stepper>.step the specific step to open can be set. This can be used to
set the stepper to the most current step on initializing, and could also
be used if the store issues a reset action. The reset method of the
stepper is not necessarily needed, as the completed state of the stepper
depends on the store, which is still automatically reset.