Capitalize parameter group names
When listing the parameter group names it looks better if the names are
capitalized. However, only German has them capitalized by default. Since
everywhere where we want to show the names we want them capitalized we
can do it directly in the parameter service when teh parameter groups
are read.