Parse assets without time-range #75
Monthly and yearly assets do not have a distinction between now and
recent, which is why they omit the time-range part fully in the file
names. They always contain all data. To a user this is basically the
same as historical assets. This is why these assets are treated the same
as historical assets without a date-range.