Parse windows-1252 encoded CSV #83
The users of the STAC api are likely to open the downloaded CSVs in
Excel. Excel is not able to handle UTF-8 CSV-files correctly if they
don't include the BOM. Currently it is unfortunately not possible to
create the CSV with the UTF8-BOM, so in order for them to work in Excel
without encoding issues they are encoded in windows-1252. This is an
issue in JavaScript, as it expects all data to be UTF-8 encoded, as
defined by the HTML5 standard. Special characters are then showed as a
black question mark.

In order to mitigate this, the CSVs are first decoded using the Encoding
API [0]. For this we need to read the bytes from the request, to make
sure the encoding is preserved. Papaparse does not support this, so we
have to fetch the data ourselves first, decode it and then provide it as
a string to papaparse.

Currently, the encoding is hard-coded in the stat-client-config as
windows-1252. There are ways to autodetect the encoding in Javascript in
a browser, but due to time constraints it makes more sense to do that
later when the STAC-api changes.

[0]: https://developer.mozilla.org/en-US/docs/Web/API/Encoding_API