Change formate for historical date ranges #41
The format for historical date ranges was changed to `YYYY-YYYY` in the
csv filename. For this the parsing had to be adjusted.

We opted to keep the date objects in place. There is a chance that month
and day will be reintroduced at a later stage so we want to keep the
option for that.

Some tests were changed to not test for correct date transformation
anymore. These tests test functions which rely on functions from
`date-transformation.utils.ts`. These functions are already tested and
test that transforming string to date and back works. This means that
there is no need to test this again in the other tests and can be
assumed working. This reduces the number of test that need to be
adjusted when the format next changes.