Steps show the currently selected value #89
The new `step-label.component` contains all logic to display the current
step label (duh!). Currently, this means that the title can be followed
by the selected value for all steps that have a selectable value.
Also it's possible to show a description. And as a special case: The
description can also contain an info tooltip icon with more description
text.