Initialize app with latest step possible #116
The latest step gets automatically selected after a debounce time of
100ms; otherwise multiple steps would be selected depending on how long
they took to load. This is done using a new store property
`initial step` that gets selected after initializing the application
using URL parameters. The logic is done completely within the forms
reducer and is basically:
Go to the review step if all values are set. Or go to the time range
step if only the time range is missing. And so on.

Also fixes the automatic selection of collections if only the station
was given in the URL but no collection. Now the collection gets
automatically selected (if no collection is already selected).