Replace all enums with constants #116
Also changed the routes to include a default route for unknown pages
back to the main page (as we don't have a 404 page). Also remove the
"multi language" TODOs because of two reasons: first, it's not easy to
add translation at that moment and secondly, this application is always
within an iframe and the page titles therefore are not important.