Fix asset loading state not set correctly #93
When a collection is selected, the assets of the stations in the
selected collection need to be retrieved from stac. This should be
represented in the store by having the loadingState set to 'loading'.
This was not done previously as the effect that reacted on the
collection selection also called the service and the reducer to set the
loading state was not involved.