Add indicator to communicate startup loading #93
This indicator is used at the startup of the application where the
metadata-csvs and, if applicable, the station assets need to be loaded
from STAC. Once the initial data is loaded and the application is ready
to take interactions from the user, the indicator is removed.

This commit also changes how the effect
initializeSelectedDataIntervalAndTimeRange works. It now always
dispatches the action initializeSelectedDataIntervalAndTimeRange even if
the stationIdd or collectionId is null. In that case it dispatches the
action with null values. This is done to make sure the value
isDataIntervalAndTimeRangeInitialized is always set in the store so
things like the loading indicator or setting the current step in the
stepper can react on that value to know when initialization is finished.