Add loading indicator in interval-selection #93
The interval selection step can only be done, if the available assets
have been loaded for the selected station and collection. These indicate
which selection options are available to the user. The fetch happens
immediately after a collection is selected. But depending on the network
speed there might be a case where the user clicks the next button but
the data is not yet available. In this case it should be communicated to
the user using a loading indicator in the step, that the selection is
not yet ready. Once the data is loaded the indicator is replaced with
the selection.

When adding the `@if` directive to display the loading indicator a
angular issue occurred where some value was changed after it was
checked. This happened for some reason in the collection-selection step.
This could be fixed by encapsulating the radio-buttons in a
radio-button-group and setting the corresponding values. I'm not sure
why that happened now, though.